// interface/enums/itemType.ts
export enum ItemType {
    WEAPON = '무기',
    JACKET = '상의',
    SHOULDER = '머리어깨',
    PANTS = '하의',
    SHOES = '신발',
    WAIST = '허리',
    AMULET = '목걸이',
    WRIST = '팔찌',
    RING = '반지',
    SUPPORT = '보조장비',
    MAGIC_STON = '마법석',
    EARRING = '귀걸이'
}